requirements.txt
# Deep Learning Framework
tensorflow>=2.10.0
keras>=2.10.0

# API Framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6

# Image Processing
pillow>=10.0.0
opencv-python>=4.8.0

# Numerical Computing
numpy>=1.23.0

# Additional Dependencies
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4